discard gui.js
