+++ js/guiManager.js
@@ -2,8 +2,7 @@ import Noodle from "./noodle.js";
 import BallAndStick from "./ballAndStick.js";
 import { numberFormatter, zIndexPanelUnselected, zIndexPanelSelected } from './utils.js';
 import { rgb255ToThreeJSColor } from "./color.js";
-import { juiceboxPanel } from "./gui.js";
-import { eventBus, noodle, ballAndStick, sceneManager } from "./app.js";
+import { eventBus, noodle, ballAndStick, sceneManager, juiceboxPanel } from "./app.js";
 
 class GUIManager {
     constructor ({ $button, $panel }) {
