+++ js/traceSelectPanel.js
@@ -1,6 +1,6 @@
 import { clamp } from './math.js'
 import Panel from "./panel.js";
-import {hideSpinner, showSpinner} from "./gui.js";
+import {hideSpinner, showSpinner} from "./app.js";
 import { ensembleManager, eventBus } from "./app.js";
 
 let currentNumber = undefined;
